Use approximate numbers.
These should succeed under `assertAlmostEqual()`, but clearly fail since it won't go into the dicts and lists.

This code now fails with:

  ERROR: test_calculate (__main__.UnitTests)
  ----------------------------------------------------------------------
  Traceback (most recent call last):
    File "test_module.py", line 29, in test_calculate
      self.assertAlmostEqual(
    File "/usr/lib64/python3.8/unittest/case.py", line 943, in assertAlmostEqual
      diff = abs(first - second)
  TypeError: unsupported operand type(s) for -: 'dict' and 'dict'

So clearly it only succeeded because `assertAlmostEqual()` short circuited the case where the two arguments are `==` to each other.

This will be fixed on the follow up commit, which will check dicts and lists and enter them.