Introduce assertDictOrListAlmostEqual.
This helper will traverse any dicts or lists or tuples, ensure that the
structure is the same on the other side, and finally use assertAlmostEqual
after it finds numbers on both sides.

This fixes the issue of using assertAlmostEqual on a dict, which would only
succeed because there's a comparison for *exact* equality at the top of it. But
if exact equality was desired, then there's no point in using
assertAlmostEqual.

Furthermore, the error messages should be passed with an explicit `msg=`
keyword argument, since the assertAlmostEqual function actually takes `places`
as its third argument, so whenever the explicit check wouldn't work, the code
would break trying to use the string message as a `places` to compare.
(Actually, it would break before that, when trying to subtract the two dicts.)

Tested this code for success with a valid solution.

Tested it with approximate numbers in the `expected` arrays. (See preceding
commit that drops a few digits at the end of each of the floats.)

Also, importantly, tested it for specific bugs in the implementation, to verify
they produce a valid exception. Tested the following cases:

  * Produced dict doesn't contain all expected keys.

    AssertionError: Items in the second set but not the first: 'min', 'max'
    Expected different output when calling 'calculate()' with '[9,1,5,3,3,3,2,9,0]'

  * Produced dict contains additional keys other than the expected ones.

    AssertionError: Items in the first set but not the second: 'extra'
    Expected different output when calling 'calculate()' with '[9,1,5,3,3,3,2,9,0]'

  * There are more elements than expected.

    AssertionError: Sequences differ: [...] != [...]
    First sequence contains 1 additional elements.

  * Produced values are 0.00000001 above calculation (no errors produced.)

    Ran 3 tests in 0.001s
    OK

  * Produced values are 0.0000001 above calculation (error is produced.)

    AssertionError: 3.8888889888888887 != 3.8888888888888 within 7 places (1.0000008865418408e-07 difference)
    Expected different output when calling 'calculate()' with '[9,1,5,3,3,3,2,9,0]'